backend/requirements.txt
Flask==2.0.1
requests==2.25.1
validators==0.18.2
python-dotenv==0.17.1
beautifulsoup4==4.9.3
textblob==0.15.3
